% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cards.R
\name{get_card_by_code}
\alias{get_card_by_code}
\title{Get card by set code and collector number}
\usage{
get_card_by_code(
  code = "wwk",
  number = 31,
  lang = "en",
  format = "json",
  face = "front",
  version = "large",
  pretty = FALSE,
  rulings = FALSE
)
}
\arguments{
\item{code}{the 3 to 5 letter set code}

\item{number}{the collector number of the card (must be whole number)}

\item{lang}{the 2 to 3 letter language code as described at https://scryfall.com/docs/api/languages}

\item{format}{The format to return the data in. "json", "CSV", "image"}

\item{face}{The face that should be returned if format selected is "image"}

\item{version}{The size of the image to return when using "image"}

\item{pretty}{Should the JSON be prettified}

\item{rulings}{(Boolean) should the rulings for the card be returned?}
}
\description{
Get card by set code and collector number
}
