% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cards.R
\name{get_card_by_id}
\alias{get_card_by_id}
\title{Get card by ID}
\usage{
get_card_by_id(
  id = 195297,
  format = "json",
  face = "front",
  version = "large",
  pretty = FALSE,
  id_type = "multiverse"
)
}
\arguments{
\item{id}{Arena ID}

\item{format}{The format to return the data in. "json", "CSV", "image"}

\item{face}{The face that should be returned if format selected is "image"}

\item{version}{The size of the image to return when using "image"}

\item{pretty}{Should the JSON be prettified}

\item{id_type}{what kind of id is being provided See details for options.}
}
\value{
either a list, data frame, or image depending on \code{format}
}
\description{
Get a single card by it's given id from one of the following ID schemes:
\itemize{
\item multiverse
\item mtgo
\item arena
\item tcgplayer
\item cardmarket
\item scryfall
See details for information about each of these ID systems.
}
}
\details{
id_type
\itemize{
\item 'multiverse': Get a single card by given Multiverse ID. If the card has multiple Multiverse IDs this method finds both of them.
A Multiverse ID is the unique identifier from Wizards of the Coast's Gatherer database.
\item `mtgo': Get a single card by given it's MTGO ID, also known as Catalog ID. These can be either the mtgo_id or it's mtgo_foil_id.
\item 'arena': Get a single card given it's arena ID
\item 'tcgplayer:' Get a single card with the given tcgplayer_id, also known as the productId on TCGplayer’s API
\item 'cardmarket' Get a single card with the given cardmarket_id, also known as the idProduct" or the Product ID on Cardmarket’s APIs
\item 'scryfall' Get a single card with the given Scryfall ID
}
}
