% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cards.R
\name{get_card_by_multiverse_id}
\alias{get_card_by_multiverse_id}
\title{Get card by Multiverse ID}
\usage{
get_card_by_multiverse_id(
  id = 195297,
  format = "json",
  face = "front",
  version = "large",
  pretty = FALSE
)
}
\arguments{
\item{id}{Multiverse ID}

\item{format}{The format to return the data in. "json", "CSV", "image"}

\item{face}{The face that should be returned if format selected is "image"}

\item{version}{The size of the image to return when using "image"}

\item{pretty}{Should the JSON be prettified}
}
\value{
either a list, data frame, or image depending on \code{format}
}
\description{
Get a single card by given Multiverse ID. If the card has multiple Multiverse IDs this method finds both of them.
A Multiverse ID is the unique identifier from Wizards of the Coast's Gatherer database. In general it appears to be ordered as
alphabetically by set release order with from Alpha starting at 94. It's likely that this is an incrementing integer key, with
A few oddities.
}
