% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cards.R
\name{get_cards_autocomplete}
\alias{get_cards_autocomplete}
\title{Get Potential Card names}
\usage{
get_cards_autocomplete(q = "Jace, ", pretty = FALSE, include_extras = FALSE)
}
\arguments{
\item{q}{The query string}

\item{pretty}{Should the JSON be prettified}

\item{include_extras}{Should the search include extra cards (tokens, planes, vanguards, etc)}
}
\value{
A list
}
\description{
Returns up to 20 full English card names that could
autocomplete the supplied query string.
}
