% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cards.R
\name{get_cards_by_search}
\alias{get_cards_by_search}
\title{Search for cards on Scryfall}
\usage{
get_cards_by_search()
}
\arguments{
\item{q}{the string query using Scryfall Syntax}

\item{unique}{the rollup mode for removing duplicates. See details}

\item{order}{the method to sort returned cards. See details}
}
\value{
A list
}
\description{
Get a list of cards found using a fulltext search string.
This string uses the same syntax as \href{https://scryfall.com/docs/syntax}{Scryfall's site}
}
\details{
\code{unique} unique parameter specifies if Scryfall should remove “duplicate” results in your query. The options are:
\itemize{
\item "cards" (default): Removes duplicate gameplay objects (cards that share a name and have the same functionality).
\item "art" : Returns only one copy of each unique artwork for matching cards.
\item "prints": Returns all prints for all cards matched (disables rollup).
}
}
