% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cards.R
\name{get_cards_random}
\alias{get_cards_random}
\title{Get  single random Card object}
\usage{
get_cards_random(
  q = "Jace",
  format = "csv",
  face = "front",
  version = "large",
  pretty = FALSE
)
}
\arguments{
\item{q}{The query string}

\item{format}{The format to return the data in. "json", "CSV", "image"}

\item{face}{The face that should be returned if format selected is "image"}

\item{version}{The size of the image to return when using "image"}

\item{pretty}{Should the JSON be prettified}
}
\description{
Returns a single random Card object.
}
