% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalogs.R
\name{get_catalog}
\alias{get_catalog}
\title{Get a Scryfall Catalog}
\usage{
get_catalog(catalog = "card-names", pretty = FALSE)
}
\arguments{
\item{catalog}{Name of the catalog to fetch}

\item{pretty}{Should the response be returned in the pretty JSON format}
}
\value{
A list
}
\description{
A Catalog object contains an array of Magic datapoints
(words, card values, etc). Catalog objects are provided by the API
as aids for building other Magic software and understanding possible
values for a field on Card objects.
}
