% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sets.R
\name{get_set_by_code}
\alias{get_set_by_code}
\title{Get set by set code}
\usage{
get_set_by_code(code = "wwk")
}
\arguments{
\item{code}{the 3 to 5 letter set code}
}
\value{
a data.frame with set details
}
\description{
Retrieve set details using the set's 3-5 character code
}
