% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sets.R
\name{get_set_by_id}
\alias{get_set_by_id}
\title{Get set by set code}
\usage{
get_set_by_id(id, id_type = "scyfall")
}
\arguments{
\item{id}{the 3 to 5 letter set code}

\item{id_type}{the type of ID provided either "scryfall" or "tcgplayer"}
}
\value{
a data.frame with set details
}
\description{
Retrieve set details using the set's 3-5 character code
}
