% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sets.R
\name{get_sets}
\alias{get_sets}
\title{Get a list of all sets}
\usage{
get_sets()
}
\value{
a data.frame containing all listed sets
}
\description{
Fetch a list of all sets, and their respective codes/IDs, release date, and set attributes.
}
