% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_if_not_in}
\alias{stop_if_not_in}
\title{stop if not in}
\usage{
stop_if_not_in(var, list, msg)
}
\arguments{
\item{var}{Name of var to check}

\item{list}{list of acceptable options}

\item{msg}{the warning message to display}
}
\description{
stop if not in
}
